requirements.txt
openai>=1.54.0
rich>=13.9.0
PyYAML>=6.0.2
dashscope>=1.20.14
Jinja2>=3.1.4
pytest>=8.3.3
pytest-asyncio>=0.24.0
